cross-intent-forecast-solution/requirements.txt
albumentations==1.4.24
hydra-core==1.3.2
lightning==2.5.1
numpy==2.1.3
onnx==1.17.0
onnxruntime-gpu==1.20.1
opencv-python-headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
pyarrow==18.1.0
pytest==8.3.4
pytest-cov==6.0.0
pytorch-lightning==2.5.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.14.1
tensorrt-cu12-bindings==10.6.0; platform_system == "Linux"
tensorrt-cu12-libs==10.6.0; platform_system == "Linux"
timm==1.0.13
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
wandb==0.19.1
